Style: 설교 홈 콘텐츠 리디자인 (Featured 히어로·최근 설교 리스트)
- Featured: 흰 카드 좌우 분할 → 풀블리드 오버레이 scrim 히어로(pill·play·하단 구절·제목·meta)
- 최근 설교: 가로 캐러셀 → 세로 리스트 가로형 카드, SermonRecentList로 rename하고 서버 컴포넌트화
