+++ src/app/(content)/sermons/_component/SermonFeatured/SermonFeatured.tsx
@@ -17,51 +17,41 @@ export default function SermonFeatured({ sermon }: Props) {
   const thumb = cloudinaryFetchUrl(getSermonThumbnail(sermon));
   const duration = formatSermonDuration(sermon.duration);
   const preacherLabel = formatPreacherLabel(sermon.preacher);
-  const summaryFirstParagraph = sermon.summary?.split('\n\n')[0] ?? '';
-  const seriesOrderLabel =
-    sermon.sermon_series && sermon.series_order !== null
-      ? String(sermon.series_order).padStart(2, '0')
-      : null;
 
   return (
     <section className={styles.section}>
       <h2 className={styles.section_title}>이번 주 설교</h2>
       <Link href={`/sermons/${sermon.id}`} className={styles.card}>
-        <div className={styles.media}>
         {thumb && (
           <CloudinaryImage
             src={thumb}
             alt={sermon.title}
             fill
-            sizes="(min-width: 768px) 580px, 100vw"
+            sizes="(min-width: 768px) 720px, 100vw"
             className={styles.thumb}
             priority
           />
         )}
+        <div className={styles.pills}>
+          <span className={styles.pill_latest}>최신 설교</span>
+          {sermon.sermon_series && (
+            <span className={styles.pill_series}>{sermon.sermon_series.title}</span>
+          )}
+        </div>
         <span className={styles.play_circle} aria-hidden>
           <IoPlay />
         </span>
         {duration && <span className={styles.duration}>{duration}</span>}
-      </div>
-      <div className={styles.body}>
-        <div className={styles.meta}>
-          {sermon.sermon_series && (
-            <>
-              <span className={styles.series}>
-                {sermon.sermon_series.title}
-                {seriesOrderLabel && ` · ${seriesOrderLabel}`}
-              </span>
-              <span className={styles.dot}>·</span>
-            </>
-          )}
-          <span className={styles.date}>{formattedDate(sermon.sermon_date, 'YYYY.MM.DD')}</span>
-          <span className={styles.dot}>·</span>
-          <span>{sermon.service_type}</span>
-        </div>
-        <h3 className={styles.title}>{sermon.title}</h3>
-        {sermon.scripture && <p className={styles.scripture}>{sermon.scripture}</p>}
-        {summaryFirstParagraph && <p className={styles.summary}>{summaryFirstParagraph}</p>}
-        <p className={styles.preacher}>{preacherLabel}</p>
+        <div className={styles.scrim}>
+          {sermon.scripture && <p className={styles.scripture}>{sermon.scripture}</p>}
+          <h3 className={styles.title}>{sermon.title}</h3>
+          <div className={styles.meta}>
+            <span className={styles.preacher}>{preacherLabel}</span>
+            <span className={styles.dot} aria-hidden>·</span>
+            <span>{formattedDate(sermon.sermon_date, 'YYYY.MM.DD')}</span>
+            <span className={styles.dot} aria-hidden>·</span>
+            <span>{sermon.service_type}</span>
+          </div>
         </div>
       </Link>
     </section>
