+++ src/app/(content)/sermons/_component/SermonRecentList/SermonRecentList.module.scss
@@ -1,9 +1,9 @@
-// 캐러셀 카드 폭 — mockup `SermonCarouselCard` PC 240px / 모바일 210px
-$card-width-mobile: 21rem;
-$card-width-pc: 24rem;
+// 가로형 리스트 카드 썸네일 폭 — 참조 시안 SermonListCard(모바일 124px, PC 확대)
+$thumb-width-mobile: 13rem;
+$thumb-width-pc: 18rem;
 // play 버튼은 sermon-play-hint-sm 믹스인 (단일 소스)
-// 영상 우하단 duration overlay — 토큰 미존재 (mockup `rgba(0,0,0,.7)`)
-$duration-overlay-bg: rgba(0, 0, 0, 0.7);
+// 영상 우하단 duration overlay — 토큰 미존재 (참조 `rgba(0,0,0,.5)`)
+$duration-overlay-bg: rgba(0, 0, 0, 0.5);
 
 .section {
   display: flex;
@@ -26,13 +26,8 @@ $duration-overlay-bg: rgba(0, 0, 0, 0.7);
   letter-spacing: $letter-spacing-body;
 }
 
-.header_actions {
-  display: inline-flex;
-  align-items: center;
-  gap: $spacing-12;
-}
-
 .more_link {
+  flex: 0 0 auto;
   font-size: $font-size-13;
   font-weight: $font-weight-semibold;
   color: $primary;
@@ -41,39 +36,50 @@ $duration-overlay-bg: rgba(0, 0, 0, 0.7);
   @include hover-color-shift($primary-hover);
 }
 
-.card {
-  flex: 0 0 $card-width-mobile;
+.list {
   display: flex;
   flex-direction: column;
-  background-color: $bg-card;
-  border: 1px solid $border-card;
-  border-radius: $radius-s;
-  overflow: hidden;
-  text-decoration: none;
-  color: inherit;
-  // PC drag 시 anchor ghost image · 텍스트 selection 차단 (의사결정 로그 D5)
-  user-select: none;
-  -webkit-user-drag: none;
+  gap: $spacing-8;
+  margin: 0;
+  padding: 0;
+  list-style: none;
 
   @include respond-up($breakpoint-tablet) {
-    flex-basis: $card-width-pc;
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: $spacing-12 $spacing-20;
   }
+}
+
+.card {
+  display: flex;
+  gap: $spacing-12;
+  padding: $spacing-8;
+  border-radius: $radius-s;
+  text-decoration: none;
+  color: inherit;
+
+  @include hover-bg-shift($bg-hover);
 }
 
 .thumb_box {
   position: relative;
-  aspect-ratio: 16 / 9;
+  flex: 0 0 $thumb-width-mobile;
+  aspect-ratio: 16 / 10;
+  border-radius: $radius-s;
   background: $overlay-image;
+  overflow: hidden;
   display: flex;
   align-items: center;
   justify-content: center;
-  overflow: hidden;
+
+  @include respond-up($breakpoint-tablet) {
+    flex-basis: $thumb-width-pc;
+  }
 }
 
 .thumb {
   object-fit: cover;
-  -webkit-user-drag: none;
-  pointer-events: none;
 }
 
 .play_circle {
@@ -104,13 +110,13 @@ $duration-overlay-bg: rgba(0, 0, 0, 0.7);
 .body {
   display: flex;
   flex-direction: column;
-  padding: $spacing-12 $spacing-12 $spacing-12;
+  justify-content: center;
+  gap: $spacing-4;
+  min-width: 0;
 }
 
 .label {
-  margin-bottom: $spacing-6;
   font-size: $font-size-11;
-  text-transform: none;
   overflow: hidden;
   text-overflow: ellipsis;
   white-space: nowrap;
@@ -127,33 +133,25 @@ $duration-overlay-bg: rgba(0, 0, 0, 0.7);
 }
 
 .title {
-  margin: 0 0 $spacing-2;
+  margin: 0;
   display: -webkit-box;
   -webkit-line-clamp: 2;
   -webkit-box-orient: vertical;
   overflow: hidden;
-  min-height: 2.8em;
-  font-size: $font-size-13;
-  font-weight: $font-weight-semibold;
+  font-size: $font-size-14;
+  font-weight: $font-weight-bold;
   color: $txt-primary;
   line-height: $line-height-snug;
   letter-spacing: $letter-spacing-body;
   word-break: keep-all;
 }
 
-.scripture {
-  margin: 0 0 $spacing-8;
-  font-size: $font-size-11;
-  font-weight: $font-weight-semibold;
-  color: $primary;
-}
-
 .meta_bar {
   display: flex;
+  flex-wrap: wrap;
   align-items: center;
   gap: $spacing-4;
-  padding-top: $spacing-8;
-  border-top: 1px solid $border-subtle;
+  margin-top: $spacing-2;
   font-size: $font-size-11;
   color: $txt-tertiary;
   font-variant-numeric: tabular-nums;
